pyproject.toml
[project]
name = "trading-agents"
version = "0.1.0"
description = "Multi-agent autonomous paper-trading system with LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-dotenv>=1.0.1",
  "pydantic>=2.9.0",
  "langgraph>=0.2.0",
  "langchain-core>=0.3.0",
  "httpx>=0.27.0",
  "redis>=5.0.0",
  "sqlalchemy>=2.0.0",
  "aiosqlite>=0.20.0",
  "pandas>=2.2.0",
  "numpy>=2.0.0",
  "feedparser>=6.0.11",
  "alpaca-py>=0.34.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "jinja2>=3.1.4",
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["evals"]
